refactoring subscriptable references.
Note: still need to handle job resolution in the over all system flow. For now refactored addjob to add the job to a 'scheduledStore' to give visibility on the jobs that were scheduled by any particular run. These processes are incomplete as of now, but the system is in a better state for more rigorous testing and improvement.

TODO: Future additions to the system can handle job executions correctly, and expand the functionality of the flow of this system to more realistically let jobs run etc.

Add init files for different modules; refactor uplink to resolve overloading design; 2 second spacing for the event generator; initial version for the sat_comm_service, to be matured later; clarified design and responsibility of the scheduler service for ground station emulation only. Also including mor mature error handling; mission runner to be mature in next efforts; generator test to use current time; more test coverage for scheduling engine test; simulation engine test paused to match later maturation of mission runner